refactor(doctor): extract doctrine health render helpers
