+++ tests/specify_cli/test_doctor_doctrine.py
@@ -299,7 +299,7 @@ def test_doctor_doctrine_human_renders_degraded_pack_and_invalid_profiles(
     original = doctor_mod.console
     doctor_mod.console = Console(file=buf, highlight=False, markup=True, width=200)
     try:
-        doctor_mod._render_doctrine_pack(entry, 0)
+        doctor_mod._render_doctrine_pack(doctor_mod.console, entry, 0)
     finally:
         doctor_mod.console = original
     output = buf.getvalue()
